orchestrator/email_coordinator/email_reader/gmail_reader/requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
pyyaml>=6.0.0
pytest>=7.0.0
